slightly refactor storages
