+++ annet/adapters/netbox/common/query.py
