+++ annet/adapters/netbox/v37/storage.py
@@ -8,8 +8,8 @@ from annet.adapters.netbox.common import models
 from annet.adapters.netbox.common.manufacturer import (
     is_supported, get_hw, get_breed,
 )
+from annet.adapters.netbox.common.query import NetboxQuery
 from annet.adapters.netbox.common.storage_opts import NetboxStorageOpts
-from annet.adapters.netbox.query import NetboxQuery
 from annet.annlib.netdev.views.hardware import HardwareView
 from annet.storage import Storage
 from . import api_models
@@ -22,13 +22,28 @@ logger = getLogger(__name__)
     link(P[api_models.Device].name, P[models.NetboxDevice].hostname),
     link(P[api_models.Device].name, P[models.NetboxDevice].fqdn),
 ])
-def extend_device(
+def extend_device_base(
         device: api_models.Device,
         interfaces: List[models.Interface],
         hw: Optional[HardwareView],
         breed: str,
 ) -> models.NetboxDevice:
     ...
+
+
+def extend_device(device: api_models.Device) -> models.NetboxDevice:
+    return extend_device_base(
+        device=device,
+        interfaces=[],
+        breed=get_breed(
+            device.device_type.manufacturer.name,
+            device.device_type.model,
+        ),
+        hw=get_hw(
+            device.device_type.manufacturer.name,
+            device.device_type.model,
+        ),
+    )
 
 
 @impl_converter
@@ -66,29 +81,26 @@ class NetboxStorageV37(Storage):
             **kwargs,
     ) -> List[models.NetboxDevice]:
         device_ids = {
-            device.id: extend_device(
-                device=device,
-                interfaces=[],
-                breed=get_breed(
-                    device.device_type.manufacturer.name,
-                    device.device_type.model,
-                ),
-                hw=get_hw(
-                    device.device_type.manufacturer.name,
-                    device.device_type.model,
-                ),
-            )
+            device.id: extend_device(device)
+            for device in self._load_devices(query)
+        }
+        if not device_ids:
+            return []
+
+        interfaces = self._load_interfaces(list(device_ids))
+        for interface in interfaces:
+            device_ids[interface.device.id].interfaces.append(interface)
+        return list(device_ids.values())
+
+    def _load_devices(self, query: NetboxQuery) -> List[api_models.Device]:
+        return [
+            device
             for device in self.netbox.all_devices(
                 name__ic=query.globs,
             ).results
             if is_supported(device.device_type.manufacturer.name)
             if _match_query(query, device)
-        }
-        if device_ids:
-            interfaces = self._load_interfaces(list(device_ids))
-            for interface in interfaces:
-                device_ids[interface.device.id].interfaces.append(interface)
-        return list(device_ids.values())
+        ]
 
     def _load_interfaces(self, device_ids: List[int]) -> List[
         models.Interface]:
@@ -108,18 +120,9 @@ class NetboxStorageV37(Storage):
             **kwargs,
     ) -> models.NetboxDevice:
         device = self.netbox.get_device(obj_id)
-        return extend_device(
-            device=device,
-            interfaces=self._load_interfaces([device.id]),
-            breed=get_breed(
-                device.device_type.manufacturer.name,
-                device.device_type.model,
-            ),
-            hw=get_hw(
-                device.device_type.manufacturer.name,
-                device.device_type.model,
-            ),
-        )
+        res = extend_device(device)
+        res.interfaces = self._load_interfaces([device.id])
+        return device
 
     def flush_perf(self):
         pass
